requirements.txt
flask==2.0.2
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-cors==3.0.10
openai==0.27.0
python-dotenv==0.19.2
gunicorn==20.1.0 